// src/app/models/egress.model.ts
export interface Egress {
  Id_Egreso?: string;
  Id_Proveedor?: string;
  Id_Servicio?: string;
  NombreProveedor?: string;
  NombreServicio?: string;
  Monto?: string;
  Periodo?: string;
  HechoPor?: string;
  RevisadoPor?: string;
  AutorizadoPor?: string;
  NumeroCheque?: string;
  FechaRegistro?: string;
  HoraRegistro?: string;
  Variable?: string;
  Estado?: string;
  FechaCobrado?: string;
  Id_Condominio?;
  error?;
  datePeriodo?;
  Mes?;
  Año?;
  year?;
  Saldo?;
}